Switched apt-get for apt and Removed Unneeded UPGRADE command
In apt-get, upgrade upgrades all packages on the system, not just a
single one. The correct usage is to call install again if you wish to
upgrade to the latest version. The -q=2 flag on the apt commands reduces
verbosity and removes animated progress bars that don't look nice in
opsmop output.